Add old blog redirect
This theme wants to put posts in the `posts` directory. This
means the URL of all my old posts will change. Get around this
problem by creating redirect pages at the old URLs.

I'll have to create one for every existing blog in the old
repo, but it is easy to add by adding a `redirect` to the
page header.