Correct links to Google Sites
For the escape room blog, I pointed to some Google documents that
were "Google Site" type of documents. I just found out that you
cannot share read-only access to the document itself. You have to
publish it and provide the published link. Correct the links
to those documents to use the published site.