Add summaries
For each post, add a summary sentence.

Without that explicit summary, the theme takes the first couple of
paragraphs instead. This may or may not make a decent summary of what
the post is about. Even worse, if there are elements like images, those
may get pulled in as well and cause further problems.